use round instead of time to store metrics

+++ src/main/java/ir/mesmaeili/lba/result/SimulationMetricResult.java
@@ -13,10 +13,10 @@ import java.util.Map;
 @Getter
 @Setter
 public class SimulationMetricResult {
-    private Map<Double, Double> LBFOverTimeMap = new HashMap<>();
-    private Map<Double, Double> throughputOverTimeMap = new HashMap<>();
-    private Map<Double, Double> blockingRateOverTimeMap = new HashMap<>();
-    private Map<Double, Double> averageQueueSizeOverTimeMap = new HashMap<>();
-    private Map<Double, Double> averageResponseTimeOverTimeMap = new HashMap<>();
-    private Map<Double, Double> averageMakespanTimeOverTimeMap = new HashMap<>();
+    private Map<Integer, Double> LBFOverTimeMap = new HashMap<>();
+    private Map<Integer, Double> throughputOverTimeMap = new HashMap<>();
+    private Map<Integer, Double> blockingRateOverTimeMap = new HashMap<>();
+    private Map<Integer, Double> averageQueueSizeOverTimeMap = new HashMap<>();
+    private Map<Integer, Double> averageResponseTimeOverTimeMap = new HashMap<>();
+    private Map<Integer, Double> averageMakespanTimeOverTimeMap = new HashMap<>();
 }
